implement PR feedback, move LL dialog to UserSettings
