+++ src/settings/controllers/LazyLoadingController.js
@@ -26,41 +26,4 @@ export default class LazyLoadingController extends SettingController {
     onChange(level, roomId, newValue) {
         dis.dispatch({action: 'flush_storage_reload'});
     }
-
-    canChangeTo(level, roomId, newValue) {
-        return new Promise((resolve) => this._showReloadDialog(resolve));
-    }
-
-    _showReloadDialog(onFinished) {
-        const QuestionDialog = sdk.getComponent("dialogs.QuestionDialog");
-        Modal.createDialog(QuestionDialog, {
-            title: _t("Turn on/off lazy load members"),
-            description:
-                <div>
-             { _t("To enable or disable the lazy loading of members, " + 
-                "the current synced state needs to be cleared out. " + 
-                "This also includes your end-to-end encryption keys, " +
-                "so to keep being able to decrypt all your existing encrypted messages, " +
-                "you'll need to export your E2E room keys and import them again afterwards.") }
-                </div>,
-            button: _t("Clear sync state and reload"),
-            extraButtons: [
-                <button key="export" className="mx_Dialog_primary"
-                        onClick={this._onExportE2eKeysClicked}>
-                   { _t("Export E2E room keys") }
-                </button>,
-            ],
-            onFinished,
-        });
-    }
-
-    _onExportE2eKeysClicked() {
-        Modal.createTrackedDialogAsync('Export E2E Keys', '', (cb) => {
-            require.ensure(['../../async-components/views/dialogs/ExportE2eKeysDialog'], () => {
-                cb(require('../../async-components/views/dialogs/ExportE2eKeysDialog'));
-            }, "e2e-export");
-        }, {
-            matrixClient: MatrixClientPeg.get(),
-        });
-    }
 }
