+++ src/settings/controllers/SettingController.js
@@ -38,11 +38,6 @@ export default class SettingController {
     getValueOverride(level, roomId, calculatedValue, calculatedAtLevel) {
         return null; // no override
     }
-
-    canChangeTo(level, roomId, newValue) {
-        return Promise.resolve(true);
-    }
-
     /**
      * Called when the setting value has been changed.
      * @param {string} level The level at which the setting has been modified.
